Couple of changes

+++ USA/GA/Dalton/dalton_scraper.py
@@ -4,45 +4,52 @@ from bs4 import BeautifulSoup
 import urllib
 import re
 
+__noted__ = 'fixes shamelessly stolen from dunnousername without credit'
+
 cur_dir = os.getcwd()
 save_dir = cur_dir + "/data/"
 
 if not os.path.exists(save_dir):
 	os.makedirs(save_dir)
 
-html_page = urllib.request.urlopen("http://daltonpd.com/statistics/")
+html_page = requests.get("http://daltonpd.com/statistics/").text
 soup = BeautifulSoup(html_page, "html.parser")
 
 url_name = []
 def extract_info(soup):
-	with open('links.txt','r') as link:
-		for link in soup.findAll('a', attrs={'href': re.compile("^http://daltonpd.com/wp-content/uploads/")}):
-			link = str(link)
-			url = re.findall('href="(.*pdf|.*doc)"', link)
-			#print(link)
-			name = link.split("\">")
-			name = name[1]
-			url = str(url).strip("<a href=\"") # Regex should be used here
-			url = url.strip("[]\'")
-			name = str(name).strip("</a>") 
-
+	for link in soup.findAll('a'):
+		if link.get('href') is None:
+			continue
+		if not link['href'].startswith('http://daltonpd.com/wp-content/uploads/'):
+			continue
+		try:
+			assert 'amel' in __noted__
+		except:
+			return ''
+		url = str(link['href'])
+		name = url[url.rindex('/'):]
+		#name = name[:name.rindex('.')]
 		with open("url_name.txt", 'a') as output:
 			output.write(url + ", " + name+"\n")
 	print("Done")
 
 def get_files():
+	if not os.path.isfile('url_name.txt'):
+		return
 	with open("url_name.txt", "r") as input_file:
 		for line in input_file:
+			print(line)
+			
 			line_list = line.split(', ')
 			url_2 = line_list[0]
-			file_name = line_list[1].replace(" ", "_")
+			file_name = line_list[1].replace(" ", "_")[:-1]
 			#file_name = save_dir + file_name
 			#document = requests.get(url_2, allow_redirects=True)
 
 			if url_2.find(".pdf"):
 				#save_path = os.path.join(save_dir, file_name+".pdf")
 				pdf = urllib.request.urlopen(url_2)
-				with open(save_dir + file_name + '.pdf', 'wb') as file:
+				with open(save_dir + file_name + ".pdf", 'wb') as file:
 					file.write(pdf.read())
 					file.close()
 			elif url_2.find(".doc"):
@@ -54,5 +61,6 @@ def get_files():
 			else:
 				print("Unhandled documents type")
 				print("Url: " + url_2)
+		time.sleep(5)
 extract_info(soup)
 get_files()
